chore: refactor kube config secret columns
Refactors the Kubernetes configmap secrets columns using /columns and a props.ts.

Part of #11035.
