--- packages/renderer/src/lib/configmaps-secrets/ConfigMapSecretColumnActions.svelte
@@ -1,11 +1,0 @@
-<script lang="ts">
-import ConfigmapSecretActions from './ConfigMapSecretActions.svelte';
-import type { ConfigMapSecretUI } from './ConfigMapSecretUI';
-
-interface Props {
-  object: ConfigMapSecretUI;
-}
-let { object }: Props = $props();
-</script>
-
-<ConfigmapSecretActions configMapSecret={object} />
