+++ packages/renderer/src/lib/configmaps-secrets/columns/Actions.svelte
@@ -1,0 +1,8 @@
+<script lang="ts">
+import ConfigmapSecretActions from '../ConfigMapSecretActions.svelte';
+import type { Props } from './props';
+
+let { object }: Props = $props();
+</script>
+
+<ConfigmapSecretActions configMapSecret={object} />
